requirements.txt
reportlab==3.6.2
mecab-python3
jaconv
argparse
